Restoring run mode to a full cycle, plus improvements
Now that the debugging mess is done, I will rerun the first cycle
of 2019-09-30 to double check that we are on the right path.
Then I will finish the day and archive it to prepare for 2019-10.

assimilate.csh
>   Fixed name of filter executable in the new environment diagnostics.
>   These diagnostics may be removed soon.
>   *This and future assimilations will use the filter compiled 2020-3-2
>   instead of 2020-7-28 (post cheyenne down time).*
env_run.xml
>   Changed archiving values for this cycle, to check output carefully
>   before it is (re)moved.
>   Restore DATA_ASSIMILATION_SCRIPT to normal value.
pre_submit.csh
>   More robust test for setting DOUT_S, to prevent wrong setting
>   when ending a job in the first day of a month, for debugging.
stage_cesm_files
setup_advanced_Rean.original
>   Replace tests for the existence of .i. with .r. because
>   .i. files can be left in RUNDIR, even when restarts are gone.