Temporary changes to repack_project.csh
to archive the "2020" cpl/hist files, which have the last 6 hours
of 2019 in them.  I compressed the files manually, so that section
is commented out.
Also discovered a mistake in changing directories.
All previous runs worked correctly because do_obs_space was 'true',
which caused the cd command to be executed.